(Note: libmyth version bump, plugins need recompiled.) Warnings fix, fix random minor issues (undesired variable overriding/hiding/etc)
git-svn-id: http://svn.mythtv.org/svn/trunk@4557 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmyth/mythdialogs.cpp
@@ -1346,10 +1346,10 @@ int MythPopupBox::showButtonPopup(MythMainWindow *parent, QString title,
     popup.addLabel(message, Medium, true);
     popup.addLabel("");
 
-    for( int i = 0; i < buttonmsgs.size(); i++ )
+    for (unsigned int i = 0; i < buttonmsgs.size(); i++ )
     {
         QButton *but = popup.addButton(buttonmsgs[i]);
-        if (defvalue == i)
+        if (defvalue == (int)i)
             but->setFocus();
     }
